# backend/src/main/java/com/supportticket/api/dto/CreateCommentRequest.java
package com.supportticket.api.dto;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;

public record CreateCommentRequest(
        @NotBlank @Size(max = 2000) String body
) {
}