# backend/src/main/java/com/supportticket/api/dto/CreateTicketRequest.java
package com.supportticket.api.dto;

import com.supportticket.domain.TicketPriority;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.NotNull;
import jakarta.validation.constraints.Size;

public record CreateTicketRequest(
        @NotBlank @Size(max = 200) String title,
        @NotBlank @Size(max = 5000) String description,
        TicketPriority priority,
        @NotNull Long assigneeId
) {
}