# backend/src/main/java/com/supportticket/api/dto/UpdateTicketStatusRequest.java
package com.supportticket.api.dto;

import com.supportticket.domain.TicketStatus;
import jakarta.validation.constraints.NotNull;

public record UpdateTicketStatusRequest(
        @NotNull TicketStatus status
) {
}