# specs/001-support-ticket-management/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Support Ticket Management API
  version: 1.0.0
  description: REST API for support tickets (ADMIN-managed).
servers:
  - url: http://localhost:8080
    description: Local backend
tags:
  - name: Auth
  - name: Users
  - name: Tickets
  - name: Comments
paths:
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Authenticate and obtain JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/users:
    get:
      tags: [Users]
      summary: List users (ADMIN only); filter by role for assignee picker
      security:
        - bearerAuth: []
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [USER, ADMIN]
      responses:
        '200':
          description: User list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserSummary'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/tickets:
    get:
      tags: [Tickets]
      summary: List tickets with search, filter, pagination (ADMIN)
      security:
        - bearerAuth: []
      parameters:
        - name: keyword
          in: query
          schema:
            type: string
          description: Case-insensitive partial match on title and description
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/TicketStatus'
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
        - name: sort
          in: query
          schema:
            type: string
            example: createdAt,desc
      responses:
        '200':
          description: Paginated ticket list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketPage'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Tickets]
      summary: Create ticket (ADMIN)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTicketRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/tickets/{ticketId}:
    get:
      tags: [Tickets]
      summary: Get ticket details (ADMIN)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TicketId'
      responses:
        '200':
          description: Ticket detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketDetailResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '403':
          $ref: '#/components/responses/Forbidden'
    patch:
      tags: [Tickets]
      summary: Update ticket fields (ADMIN); does not change status
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TicketId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTicketRequest'
      responses:
        '200':
          description: Updated ticket
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/tickets/{ticketId}/status:
    patch:
      tags: [Tickets]
      summary: Transition ticket status (ADMIN)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TicketId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTicketStatusRequest'
      responses:
        '200':
          description: Status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketResponse'
        '409':
          $ref: '#/components/responses/Conflict'
        '404':
          $ref: '#/components/responses/NotFound'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/tickets/{ticketId}/comments:
    post:
      tags: [Comments]
      summary: Add comment (ADMIN)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TicketId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommentRequest'
      responses:
        '201':
          description: Comment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '403':
          $ref: '#/components/responses/Forbidden'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    TicketId:
      name: ticketId
      in: path
      required: true
      schema:
        type: integer
        format: int64
  schemas:
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
          format: password
    LoginResponse:
      type: object
      properties:
        accessToken:
          type: string
        tokenType:
          type: string
          example: Bearer
        role:
          type: string
          enum: [ADMIN, USER]
    UserSummary:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        role:
          type: string
          enum: [ADMIN, USER]
    TicketPriority:
      type: string
      enum: [P0, P1, P2]
    TicketStatus:
      type: string
      enum: [OPEN, IN_PROGRESS, RESOLVED, CLOSED, CANCELLED]
    CreateTicketRequest:
      type: object
      required: [title, description, assigneeId]
      properties:
        title:
          type: string
          maxLength: 200
        description:
          type: string
          maxLength: 5000
        priority:
          $ref: '#/components/schemas/TicketPriority'
          description: Defaults to P1 if omitted
        assigneeId:
          type: integer
          format: int64
    UpdateTicketRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 200
        description:
          type: string
          maxLength: 5000
        priority:
          $ref: '#/components/schemas/TicketPriority'
        assigneeId:
          type: integer
          format: int64
    UpdateTicketStatusRequest:
      type: object
      required: [status]
      properties:
        status:
          $ref: '#/components/schemas/TicketStatus'
    CreateCommentRequest:
      type: object
      required: [body]
      properties:
        body:
          type: string
          maxLength: 2000
    TicketResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        priority:
          $ref: '#/components/schemas/TicketPriority'
        status:
          $ref: '#/components/schemas/TicketStatus'
        assignee:
          $ref: '#/components/schemas/UserSummary'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CommentResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        body:
          type: string
        author:
          $ref: '#/components/schemas/UserSummary'
        createdAt:
          type: string
          format: date-time
    TicketDetailResponse:
      allOf:
        - $ref: '#/components/schemas/TicketResponse'
        - type: object
          properties:
            comments:
              type: array
              items:
                $ref: '#/components/schemas/CommentResponse'
    TicketPage:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/TicketResponse'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string
        fieldErrors:
          type: object
          additionalProperties:
            type: string
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Not authorized (e.g. USER role)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Invalid state transition or business conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'